move failing test to correct file
