+++ src/Features/SupportLockedProperties/UnitTest.php
@@ -3,7 +3,8 @@
 namespace Livewire\Features\SupportLockedProperties;
 
 use Livewire\Livewire;
-use Livewire\Component;
+use Livewire\Component as BaseComponent;
+use Livewire\Form;
 
 class UnitTest extends \Tests\TestCase
 {
@@ -14,7 +15,7 @@ class UnitTest extends \Tests\TestCase
             'Cannot update locked property: [count]'
         );
 
-        Livewire::test(new class extends Component {
+        Livewire::test(new class extends BaseComponent {
             #[BaseLocked]
             public $count = 1;
 
@@ -35,7 +36,7 @@ class UnitTest extends \Tests\TestCase
             'Cannot update locked property: [foo]'
         );
 
-        Livewire::test(new class extends Component {
+        Livewire::test(new class extends BaseComponent {
             #[BaseLocked]
             public $foo = ['count' => 1];
 
@@ -52,7 +53,7 @@ class UnitTest extends \Tests\TestCase
     /** @test */
     function can_update_locked_property_with_similar_name()
     {
-        Livewire::test(new class extends Component {
+        Livewire::test(new class extends BaseComponent {
             #[BaseLocked]
             public $count = 1;
 
@@ -64,5 +65,43 @@ class UnitTest extends \Tests\TestCase
         })
         ->assertSet('count2', 1)
         ->set('count2', 2);
+    }
+
+    /** @test */
+    public function it_can_updates_form_with_locked_properties()
+    {
+        Livewire::test(Component::class)
+            ->set('form.foo', 'bar')
+            ->assertSet('form.foo', 'bar')
+            ->assertOk();
+    }
+}
+    
+class SomeForm extends Form {
+    #[BaseLocked]
+    public ?string $id = null;
+    public string $foo = '';
+
+    public function init(?string $id) {
+        $this->id = $id;
+    }
+}
+
+class Component extends BaseComponent
+{
+    public SomeForm $form;
+
+    public function mount() {
+        $this->form->init('id');
+    }
+
+    public function render()
+    {
+        return <<< 'HTML'
+<div>
+
+    <input type='text' wire:model='form.foo' />
+</div>
+HTML;
     }
 }
